[core] Prune schema modules from workflow bundles
