+++ .changeset/prune-workflow-sandbox.md
@@ -1,0 +1,6 @@
+---
+'@workflow/core': patch
+'@workflow/world': patch
+---
+
+Keep schema-only World modules out of workflow VM bundles.
